@@ -1,60 +1,35 @@
-import React, { Component } from "react";
-import {
-    View,
-    Dimensions
-} from "react-native";
+import React, { useEffect, useRef, useState } from "react";
+import { View, Dimensions } from "react-native";
 
-class OrientationAwareView extends Component {
-    state = {
-        portraitStyle: {},
-        landscapeStyle: {},
-        isPortrait: true
-    }
+const OrientationAwareView = (props) => {
+  const dimensionsHandler = useRef((dimensions) => {
+    setIsPortrait(dimensions.window.width < dimensions.window.height);
+  });
+  const [isPortrait, setIsPortrait] = useState(
+    Dimensions.get("window").width < Dimensions.get("window").height
+  );
 
-    dimensionsHandler = dimensions => this.calculateDeviceOrientation(dimensions);
+  useEffect(() => {
+    Dimensions.addEventListener("change", dimensionsHandler.current);
 
-    constructor(props) {
-        super(props);
+    return () => {
+      Dimensions.removeEventListener("change", dimensionsHandler.current);
+    };
+  }, []);
 
-        let defaultStyle = {};
-        if (this.props.style) {
-            defaultStyle = this.props.style;
-        }
+  if (props.hideOnPortrait && isPortrait) return null;
+  if (props.hideOnLandscape && !isPortrait) return null;
 
-        if (this.props.portraitStyle) {
-            this.state.portraitStyle = Object.assign({}, defaultStyle, this.props.portraitStyle);
-        }
-
-        if (this.props.landscapeStyle) {
-            this.state.landscapeStyle = Object.assign({}, defaultStyle, this.props.landscapeStyle);
-        }
-
-        Dimensions.addEventListener("change", this.dimensionsHandler);
-        this.calculateDeviceOrientation();
-    }
-
-    calculateDeviceOrientation(dimensions) {
-        if (dimensions) {
-            this.setState({ isPortrait: (dimensions.window.width < dimensions.window.height) });
-        } else {
-            this.state.isPortrait = (Dimensions.get("window").width < Dimensions.get("window").height);
-        }
-    }
-
-    componentWillUnmount() {
-        Dimensions.removeEventListener("change", this.dimensionsHandler);
-    }
-
-    render() {
-        if ((this.state.isPortrait && !this.props.hideOnPortrait) || (!this.state.isPortrait && !this.props.hideOnLandscape)) {
-            return (
-                <View {...this.props} style={(this.state.isPortrait ? this.state.portraitStyle : this.state.landscapeStyle)}>
-                </View>
-            );
-        }else{
-            return null;
-        }
-    }
-}
+  return (
+    <View
+      {...props}
+      style={Object.assign(
+        {},
+        props.style,
+        isPortrait ? props.portraitStyle : props.landscapeStyle
+      )}
+    ></View>
+  );
+};
 
 export default OrientationAwareView;
